+++ src/Class/Rules/StraightFlush.hpp
@@ -15,7 +15,7 @@ class StraightFlush : public Rules {
 
         /* Getter */
         int getID() const;
-        int getScore() const;
+        float getScore() const;
         std::string getName() const;
 
         /* Setter */
